# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /printers:
    get:
      operationId: listPrinters
      tags: [printers]
      summary: List all printers
      responses:
        "200":
          description: List of printers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Printer"

    post:
      operationId: createPrinter
      tags: [printers]
      summary: Create a new printer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrinterBody"
      responses:
        "201":
          description: Created printer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Printer"
  /printers/{id}:
    get:
      operationId: getPrinter
      tags: [printers]
      summary: Get a single printer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Printer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Printer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updatePrinter
      tags: [printers]
      summary: Update a printer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePrinterBody"
      responses:
        "200":
          description: Updated printer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Printer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deletePrinter
      tags: [printers]
      summary: Delete a printer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Printer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        nozzleSize:
          type: string
          description: Nozzle diameter in mm (0.2, 0.4, 0.6, 0.8)
        nozzleType:
          type: string
          description: Nozzle material (Hardened Steel or Stainless Steel)
        notes:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - isDefault
        - createdAt
        - updatedAt
    CreatePrinterBody:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        nozzleSize:
          type: string
        nozzleType:
          type: string
        notes:
          type: string
        isDefault:
          type: boolean
      required:
        - name
    UpdatePrinterBody:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        nozzleSize:
          type: string
        nozzleType:
          type: string
        notes:
          type: string
        isDefault:
          type: boolean
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string